+++ checkout-kata/Interfaces/IDiscounter.cs
@@ -4,6 +4,11 @@ namespace checkout_kata.Interfaces
 {
     public interface IDiscounter
     {
-        double CalculateDiscount(IEnumerable<Product> products);
+        /// <summary>
+        /// Calculates the total discount for the input products.
+        /// </summary>
+        /// <param name="products"></param>
+        /// <returns></returns>
+        decimal CalculateDiscount(IEnumerable<Product> products);
     }
 }
